+++ sfc-api/src/main/java/com/xiaotao/saltedfishcloud/service/share/entity/ShareExtractorDTO.java
@@ -10,7 +10,7 @@ import javax.validation.constraints.NotNull;
 @Builder
 public class ShareExtractorDTO {
     @NotNull
-    private Integer sid;
+    private Long sid;
     @NotBlank
     private String verification;
     private String code;
